feat: extract deployment helper utilities to helper module
